tuned-ppd: Keep track of the active and base profile
Instead of querying TuneD for the active profile and
coverting it to a PPD profile using a reverse profile
map, this commit makes tuned-ppd keep track of the
currently active PPD profile. This allows for more
flexible profile mappings, as they do not have to
be injective now.

Alongside the active PPD profile, tuned-ppd now also keeps
track of the "base" PPD profile, which is restored when
all profile holds are released or when tuned-ppd is restarted.
For the latter purpose, this profile is also saved in a file.